requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
supabase==2.9.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2